pyproject.toml
[project]
name = "say-it-aloud"
version = "0.1.0"
description = ""
authors = [
    {name = "Sourav Kumar Bose",email = "sourav.kumar.bose@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio (>=5.14.0,<6.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
